Script actually works now.
Added logic to skip an existing record (will add logic to overwrite later).

Removed duplicate calls to creating DNS records.

Added error handling for when adding a record fails.

Added logic for handling apex records (need to be able to identify and group apex records of the same type to suppress what would be otherwise seen as a duplicate record).